SSDM-9369 Made the test GlobalSearchTest.testSearchWithDataSetPermIdAndDataSetFetched() and a few others pass.

+++ openbis/source/java/ch/ethz/sis/openbis/generic/server/asapi/v3/search/planner/GlobalSearchManager.java
@@ -109,7 +109,7 @@ public class GlobalSearchManager implements IGlobalSearchManager
 //            SearchDomainSearchResult searchResult = searchDomain.getSearchResult();
 //            matchingEntity.setSearchDomain(searchResult.getSearchDomain().getLabel());
 
-            matchingEntity.setScore((Double) fieldsMap.get(RANK_ALIAS));
+            matchingEntity.setScore((Float) fieldsMap.get(RANK_ALIAS));
 
             final List<PropertyMatch> matches = new ArrayList<>();
